+++ cranelift/codegen/src/isa/x64/inst/encoding/mod.rs
@@ -1,0 +1,1 @@
+pub mod rex;
